Add files via upload
This C++ program implements Dijkstra's algorithm to find the shortest paths from a source vertex in a graph. The graph is represented using an adjacency matrix, and the algorithm uses a sequential approach. The program initializes a graph with random weights and measures the execution time of the algorithm, as well as the time spent in key operations like finding the minimum distance and updating the shortest paths. Timers are added to track the performance of these steps. The program runs Dijkstra's algorithm from vertex 0 across a graph with 19,999 vertices.